% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGPfunc.R
\name{DGP}
\alias{DGP}
\title{Data Generate Process}
\usage{
DGP(N, m, case)
}
\arguments{
\item{N}{sample size N}

\item{m}{dimensions}

\item{case}{simulation cases, Table 1-4}
}
\value{
a N*m matrix
}
\description{
Data Generate Process
}
\examples{
DGP(N = 4, m = 4, case = 1)
}
