% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Snm_rcpp}
\alias{Snm_rcpp}
\title{A Rcpp version of Snm}
\usage{
Snm_rcpp(X)
}
\arguments{
\item{X}{a n*m matrix}
}
\value{
a vector contain [1]Snm and [2]p-value
}
\description{
A Rcpp version of Snm
}
\examples{
Snm_rcpp(DGP(N = 4, m = 4, case = 1))
}
