% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{T3_rcpp}
\alias{T3_rcpp}
\title{A Rcpp version of T3}
\usage{
T3_rcpp(X)
}
\arguments{
\item{X}{a n*m matrix}
}
\value{
a vector contain [1]T3 and [2]p-value
}
\description{
A Rcpp version of T3
}
\examples{
T3_rcpp(DGP(N = 4, m = 4, case = 1))
}
