% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tnm_rcpp}
\alias{Tnm_rcpp}
\title{A Rcpp version of Tnm}
\usage{
Tnm_rcpp(X)
}
\arguments{
\item{X}{a n*m matrix}
}
\value{
a vector contain [1]Tnm and [2]p-value
}
\description{
A Rcpp version of Tnm
}
\examples{
Tnm_rcpp(DGP(N = 4, m = 4, case = 1))
}
