% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Trho_rcpp}
\alias{Trho_rcpp}
\title{A Rcpp version of T_rho}
\usage{
Trho_rcpp(X)
}
\arguments{
\item{X}{a n*m matrix}
}
\value{
a vector contain [1]T_rho and [2]p-value
}
\description{
A Rcpp version of T_rho
}
\examples{
Trho_rcpp(DGP(N = 4, m = 4, case = 1))
}
